requirements.txt
# Core dependencies
numpy>=1.24.0
scipy>=1.10.0

# 3D processing
plyfile>=0.7.4
scikit-learn>=1.3.0  # For KD-tree

# Computer vision
opencv-python>=4.8.0
ultralytics>=8.0.0  # YOLOv8
Pillow>=10.0.0

# Browser automation
playwright>=1.40.0

# Video processing
ffmpeg-python>=0.2.0

# Utilities
pyyaml>=6.0
tqdm>=4.66.0

